pyproject.toml
[tool.poetry]
name = "notifications_from_sahibinden"
version = "0.1.0"
description = ""
authors = ["Arkadii Paktusov <paktusov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.2"
requests = "^2.28.1"
selenium = "^4.5.0"
packaging = "^21.3"
pymongo = "^4.3.2"
python-dotenv = "^0.21.0"
celery = {extras = ["redis"], version = "^5.2.7"}
ratelimit = "^2.2.1"
pyquery = "^1.4.3"
ipython = "^8.7.0"
pylint = "^2.15.8"
isort = "^5.11.2"
black = "^22.12.0"
python-telegram-bot = {extras = ["rate-limiter"], version = "^20.0a4"}
aiolimiter = "^1.0.0"
pytest = "^7.2.1"
sqlalchemy = "^1.4.46"
alembic = "^1.9.2"
psycopg2-binary = "^2.9.5"
curl-cffi = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.master]
extension-pkg-allow-list = ["pydantic"]

[tool.pylint.format]
max-line-length = 120


[tool.pylint."MESSAGES CONTROL"]
disable = [
  "C0103",  # invalid-name (for api routes)
  "C0114",  # missing-module-docstring
  "C0115",  # missing-class-docstring (for api routes)
  "C0116",  # missing-function-docstring
  "E0213", # no-self-argument (for pydantic classes)
  "R0902", # too-many-instance-attributes (for pydantic classes)
  "R0903",  # too-few-public-methods (for sqlalchemy tables and pydantic classes)
]

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
known_local_folder = "app bot storage"
lines_after_imports = 2
multi_line_output = 3
line_length = 120
include_trailing_comma = true
